requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pyTelegramBotAPI==4.13.0
django-environ==0.11.2
psycopg2-binary==2.9.6